+++ telegram_bot/filters/__init__.py
@@ -1,2 +1,2 @@
-from .main import LocalizedTextFilter, ReplyToSupportMessageFilter
+from .main import LocalizedTextFilter, FromBotToAdminFilter
 from .register import FullNameFilter, IsNotRegisteredFilter
