+++ telegram_bot/filters/main.py
@@ -18,11 +18,12 @@ class LocalizedTextFilter(BaseFilter):
 
 		return message.text == l10n.format_value(self.l10n_key)
 
-class ReplyToSupportMessageFilter(BaseFilter):
-    async def __call__(self, message: Message) -> bool:
-        if not message.reply_to_message:
-            return False
-        if message.chat.id != SUPPORT_CHAT_ID:
-            return False
-
-        return message.reply_to_message.text.startswith("support")
+class FromBotToAdminFilter(BaseFilter):
+	async def __call__(self, message: Message) -> bool:
+		if not message.reply_to_message:
+			return False
+		if message.chat.id != SUPPORT_CHAT_ID:
+			return False
+		if message.reply_to_message.from_user.id != message.bot.id:
+			return False
+		return True
