+++ telegram_bot/handlers/main.py
@@ -1,6 +1,6 @@
 from aiogram import Dispatcher
 
-from .admin import admin_router
+from .admin_menu import admin_main_router
 from .promocode import promocode_router
 from .user import user_router
 
@@ -9,5 +9,5 @@ def register_handlers(dp: Dispatcher):
 	dp.include_routers(
 		user_router,
 		promocode_router,
-		admin_router,
+		admin_main_router,
 	)
