+++ telegram_bot/utils/__init__.py
@@ -1,2 +1,1 @@
 from .escape import escape_md_v2
-from .support_data import SupportFactory
